refactor: split admin panel tests
